+++ packages/vexide-devices/src/peripherals.rs
@@ -27,21 +27,28 @@ use core::sync::atomic::AtomicBool;
 use crate::{
     adi::AdiPort,
     controller::{Controller, ControllerId},
-    screen::Screen,
+    display::Display,
     smart::SmartPort,
 };
 
 static PERIPHERALS_TAKEN: AtomicBool = AtomicBool::new(false);
 
-#[derive(Debug)]
-/// A struct that contains all ports on the V5 Brain
-/// and guarantees **at compile time** that each port is only used once.
-/// Because of the fact that this checks at compile time, it cannot be moved once it has been used to create a device.
+/// Contains an instance of a brain’s available I/O, including ports, hardware, and devices.
+///
+/// A brain often has many external devices attached to it. We call these devices *peripherals*, and this
+/// struct is the "gateway" to all of these. [`Peripherals`] is intended to be used as a singleton, and you
+/// will typically only get one of these in your program's execution. This guarantees **at compile time** that
+/// each port is only used once.
+///
+/// Because of the fact that this checks at compile time, it cannot be copied, cloned, or moved once
+/// it has been used to create a device.
+///
 /// If you need to store a peripherals struct for use in multiple functions, use [`DynamicPeripherals`] instead.
 /// This struct is always preferred over [`DynamicPeripherals`] when possible.
+#[derive(Debug)]
 pub struct Peripherals {
-    /// Brain screen
-    pub screen: Screen,
+    /// Brain display
+    pub display: Display,
 
     /// Primary ("Master") Controller
     pub primary_controller: Controller,
@@ -116,7 +123,7 @@ impl Peripherals {
         // SAFETY: caller must ensure that this function is only called once
         unsafe {
             Self {
-                screen: Screen::new(),
+                display: Display::new(),
 
                 primary_controller: Controller::new(ControllerId::Primary),
                 partner_controller: Controller::new(ControllerId::Partner),
@@ -187,7 +194,7 @@ impl Peripherals {
 /// When possible, use [`Peripherals`] instead.
 #[derive(Debug)]
 pub struct DynamicPeripherals {
-    screen: bool,
+    display: bool,
     smart_ports: [bool; 21],
     adi_slots: [bool; 8],
 }
@@ -201,7 +208,7 @@ impl DynamicPeripherals {
         let smart_ports = [false; 21];
         let adi_slots = [false; 8];
         Self {
-            screen: false,
+            display: false,
             smart_ports,
             adi_slots,
         }
@@ -237,13 +244,13 @@ impl DynamicPeripherals {
         Some(unsafe { AdiPort::new(port_number as u8 + 1, None) })
     }
 
-    /// Creates a [`Screen`] only if one has not been created before.
-    pub fn take_screen(&mut self) -> Option<Screen> {
-        if self.screen {
+    /// Creates a [`Display`] only if one has not been created before.
+    pub fn take_display(&mut self) -> Option<Display> {
+        if self.display {
             return None;
         }
-        self.screen = true;
-        Some(unsafe { Screen::new() })
+        self.display = true;
+        Some(unsafe { Display::new() })
     }
 }
 impl From<Peripherals> for DynamicPeripherals {