+++ CHANGELOG.md
@@ -24,14 +24,31 @@ Before releasing:
 
 ### Added
 
+- You can now detect controller release occurrences with `ButtonState::is_now_released`.
+- Added support for 5.5W motors with a new constructor (`Motor::new_exp`) and four new getters (`Motor::max_voltage`, `Motor::motor_type`, `Motor::is_v5`, and `Motor::is_exp`) for `Motor`. (#167)
+
 ### Fixed
 
+- The `dbg!();` now works as expected when no arguments are supplied to it. (#175)
+- `Motor::velocity` now correctly returns the estimated velocity instead of target velocity. (#184) (**Breaking Change**)
+
 ### Changed
+
+- Controller state is now returned all at once to reduce error checking. (#152) (**Breaking Change**)
+- `Button::was_pressed` has been renamed to `ButtonState::is_now_pressed`.
+- `battery::capacity` now returns from 0.0-1.0 rather than 0-100.
+- `battery::voltage` is now returned in volts rather than millivolts.
+- `battery::current` is now returned in amps rather than milliamps.
+- Changed the incorrect return types of `AdiSolenoid::is_open` and `AdiSolenoid::is_closed` from `LogicLevel` to `bool`. (#164) (**Breaking Change**)
+- Renamed `Motor::MAX_VOLTAGE` to `Motor::V5_MAX_VOLTAGE` and added `Motor::EXP_MAX_VOLTAGE`. (#167) (**Breaking Change**)
+- Moved the ability to convert Smart devices to `SmartPorts` out of the `SmartDevice` trait and into the devices themselves. (#171) (**Breaking Change**)
+- Made the following functions infallible: `AdiAccelerometer::sensitivity`, `AdiAccelerometer::max_acceleration`, `AdiPotentiometer::potentiometer_type`, `AdiPotentiometer::max_angle`, `Motor::target`, and `RotationSensor::direction`. (#182) (**Breaking Change**)
 
 ### Removed
 
 ### New Contributors
 
+@zabackary made their first contribution in #164!
 
 ## [0.4.1]
 
@@ -72,6 +89,7 @@ Before releasing:
 - `SmartPort::device_type` now no longer returns a `Result`. (#121) (**Breaking Change**)
 - Updated the names of certain misspelled `enum` variants, constants, and fields. (#132) (**Breaking Change**)
 - Marks many futures as `#[must_use]` to warn when futures are created without `await`ing them. (#112)
+- Renamed `Screen` and its associated structs to `Display`. (#138) (**Breaking Change**)
 - Changes the banner attribute syntax in the `vexide::main` macro. (#127) (**Breaking Change**)
 - Controller joystick axis getters now return `f64` instead of `f32`. (#133) (**Breaking Change**)
 - Fixed an issue where the async executor would block indefinetly on the first program run after a Brain reboot (#139)